Add files via upload
Note, the "cav" file imported in merge_path_2 is renamed to be cav_for_merge_path for this repository

merge_path_2 controls limo770 to correctly navigate the merging path built in the RASTIC arena. By changing the coordinates in cav.py, the map can be reshaped and moved elsewhere. Changing other parameters in generate_line() allows for further customization and tuning, ensuring that the virtual map results in an accurate depiction of the physical map